# tomp_quickstart_oi/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Transport Operator MaaS Provider API
  description: "An API between MaaS providers and transport operators for booking\
    \ trips and corresponding assets. <p>The documentation (examples, process flows\
    \ and sequence diagrams) can be found at <a href=\"https://github.com/TOMP-WG/TOMP-API/\"\
    >github</a>."
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.3.0
servers:
- url: /
...
paths:
...
  /operator/meta:
    get:
      tags:
      - operator information
      - TO
      - MP
      summary: describes the running implementations
      description: "all versions that are implemented on this url, are described in\
        \ the result of this endpoint. In contains all versions and per version the\
        \ endpoints, their status and the supported scenarios."
      operationId: operator_meta_get
      parameters:
      - name: Accept_Language
        in: header
        description: "A list of the languages/localizations the user would like to\
          \ see the results in. For user privacy and ease of use on the TO side, this\
          \ list should be kept as short as possible, ideally just one language tag\
          \ from the list in operator/information"
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: A comma-separated list of BCP 47 (RFC 5646) language tags and optional
            weights as described in IETF RFC7231 section 5.3.5
        example: "nl, de;q=0.7"
      - name: maas_id
        in: header
        description: The ID of the sending maas operator
        required: true
        style: simple
        explode: false
        schema:
          type: string
        example: 1324A-DFB3482-32ACD
      - name: addressed_to
        in: header
        description: The ID of the maas operator that has to receive this message
        required: false
        style: simple
        explode: false
        schema:
          type: string
        example: 1324A-DFB3482-32ACD
      responses:
        "200":
          description: successful operation
          headers:
            Content-Language:
              description: The language/localization of user-facing content
              required: true
              style: simple
              explode: false
              schema:
                type: string
                format: One IETF BCP 47 (RFC 5646) language tag
              example: nl
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/endpointImplementation'
                x-content-type: application/json
      x-openapi-router-controller: swagger_server.controllers.operator_information_controller
...
components:
  schemas:
...
    endpoint:
      required:
      - method
      - path
      - status
      type: object
      properties:
        method:
          type: string
          enum:
          - POST
          - PUT
          - GET
          - DELETE
          - PATCH
        path:
          type: string
          description: "the exact path of the endpoint, starting after the base URL"
          example: /plannings/
        eventType:
          type: string
          description: "in case the path is ending in /events, the event type/operator\
            \ enum should be added here."
          enum:
          - PREPARE
          - ASSIGN_ASSET
          - SET_IN_USE
          - PAUSE
          - START_FINISHING
          - FINISH
          - ISSUE
          - CANCEL
          - EXPIRE
          - DENY
          - COMMIT
        status:
          type: string
          enum:
          - NOT_IMPLEMENTED
          - DIALECT
          - IMPLEMENTED
        supportsPaging:
          type: boolean
          description: does this endpoint support paging? In that case this endpoint
            can be accessed using query parameters offset=x and limit=y. Only allowed
            at endpoints that have specified these query parameters.
          default: false
        maxPageSize:
          type: number
          description: "the maximum size of the pages (only valid when supportsPaging=true).\
            \ If the limit-parameter of the request is above this amount, a http code\
            \ 400 will be returned."
      description: a formal description of an endpoint.
      example:
        withoutPaging:
          method: POST
          path: "/booking/{id}/events"
          eventType: COMMIT
          status: IMPLEMENTED
        withPaging:
          method: POST
          path: /operator/stations
          status: IMPLEMENTED
          supportsPaging: true
          maxPageSize: 100
    endpointImplementation:
      required:
      - baseUrl
      - endpoints
      - processIdentifiers
      - scenarios
      - version
      type: object
      properties:
        version:
          type: string
        baseUrl:
          type: string
        endpoints:
          type: array
          items:
            $ref: '#/components/schemas/endpoint'
        scenarios:
          type: array
          items:
            $ref: '#/components/schemas/scenario'
        processIdentifiers:
          $ref: '#/components/schemas/processIdentifiers'
      description: "a complete endpoint description, containing all endpoints, their\
        \ status, but also the served scenarios and implemented process flows. The\
        \ identifiers for the process flows can be found at https://github.com/TOMP-WG/TOMP-API/wiki/ProcessIdentifiers"
      example:
        version: 0.5.0
        baseUrl: https://dummy-bikes.org/
        endpoints: []
        scenarios:
        - POSTPONED_COMMIT
        processIdentifiers:
          planning:
          - LOCATION_BASED
          booking:
          - ACCESS_CODE_QR
          - ACCESS_CODE_IN_COMMIT_EVENT
...
    processIdentifiers:
      required:
      - booking
      - general
      - operatorInformation
      - payment
      - planning
      - support
      - tripExecution
      type: object
      properties:
        operatorInformation:
          type: array
          items:
            type: string
        planning:
          type: array
          items:
            type: string
        booking:
          type: array
          items:
            type: string
        tripExecution:
          type: array
          items:
            type: string
        support:
          type: array
          items:
            type: string
        payment:
          type: array
          items:
            type: string
        general:
          type: array
          items:
            type: string
...
    scenario:
      type: string
      enum:
      - POSTPONED_COMMIT
      - DEPOSIT
      - PAY_WHEN_FINISHED
      - REQUIRE_BOOKING_DATA
      - RETURN_AREA
      - UPFRONT_PAYMENT
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      x-basicInfoFunc: swagger_server.controllers.authorization_controller.check_BasicAuth
    BearerAuth:
      type: http
      scheme: bearer
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_BearerAuth
    ApiKeyAuth:
      type: apiKey
      name: X-API-Key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_ApiKeyAuth
    OAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes: {}
      x-tokenInfoFunc: swagger_server.controllers.authorization_controller.check_OAuth
      x-scopeValidateFunc: swagger_server.controllers.authorization_controller.validate_scope_OAuth
    OpenId:
      type: openIdConnect
      openIdConnectUrl: https://unknownserver/.well-known/openid-configuration